Refactor backend for improved structure and coherence
- Consolidated routing into a central router to activate all route files.
- Modularized controllers, separating logic into `air`, `ai`, `utility`, and `root` controllers.
- Removed the redundant `simple.controller.js` and `simple.routes.js`.
- Moved WebSocket connection logic to a dedicated `websocket/realtime.js` module.
- Extracted helper functions into new `utils` files.
- Added a unit test for the `/health` endpoint and configured Jest for ES module support.
